Reorg logging and metrics code into shared/monitoring directory
