+++ backend/api/src/serve.ts
@@ -1,7 +1,7 @@
 import * as admin from 'firebase-admin'
 import { getLocalEnv, initAdmin } from 'shared/init-admin'
 import { loadSecretsToEnv, getServiceAccountCredentials } from 'common/secrets'
-import { log } from 'shared/log'
+import { log } from 'shared/utils'
 
 const LOCAL_DEV = process.env.GOOGLE_CLOUD_PROJECT == null
 if (LOCAL_DEV) {
