+++ backend/shared/src/bounty.ts
@@ -1,10 +1,9 @@
 import * as admin from 'firebase-admin'
 
 import { runAwardBountyTxn } from './txn/run-bounty-txn'
-import { log } from './log'
 import { createSupabaseDirectClient } from './supabase/init'
 import { updateData } from './supabase/utils'
-import { getUser } from 'shared/utils'
+import { getUser, log } from 'shared/utils'
 import { APIError } from 'common/api/utils'
 import { getUserPortfolioInternal } from 'shared/get-user-portfolio-internal'
 import { canSendMana } from 'common/can-send-mana'
