+++ backend/shared/src/utils.ts
@@ -26,8 +26,8 @@ import { convertUser } from 'common/supabase/users'
 import { convertContract } from 'common/supabase/contracts'
 import { Row } from 'common/supabase/utils'
 import { SafeBulkWriter } from 'shared/safe-bulk-writer'
-import { log, Logger } from 'shared/log'
-import { metrics } from 'shared/gcp-metrics'
+import { log, Logger } from 'shared/monitoring/log'
+import { metrics } from 'shared/monitoring/metrics'
 
 export { metrics }
 export { log, Logger }
